+++ CHANGELOG.md
@@ -3,5 +3,12 @@
 # find-immutables-usages-plugin Changelog
 
 ## [Unreleased]
+
+### Added
+- Enhance 'Find Usages' action on property getter in interface/abstract class
+  that has `@Value.Immutable` to look also for builder methods and `with*()` methods
+  that provide value for that property
+
+## 0.0.1
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)